Explicitely cast to double to make our intent more clear.

+++ source/tools/matchbin_metrics.h
@@ -100,7 +100,7 @@ namespace emp {
       return static_cast<double>(emp::hash_combine(
         qhasher(a),
         thasher(b)
-      )) / std::numeric_limits<size_t>::max();
+      )) / (static_cast<double>(std::numeric_limits<size_t>::max()));
     }
 
   };
